Tests/LibWeb/Text/input/HTML/HTMLMediaElement-addTextTrack.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    test(() => {
        const video = document.createElement("video");
        video.textTracks.addEventListener("addtrack", () => {
            println(`addtrack event called`);
        });

        const track = video.addTextTrack("subtitles", "demo label", "en-GB");
        println(`track.kind: ${track.kind}`);
        println(`track.label: ${track.label}`);
        println(`track.language: ${track.language}`);
        println(`track.mode: ${track.mode}`);
    });
</script>
