Tests/LibWeb/Text/expected/HTML/HTMLMediaElement-addTextTrack.txt

track.kind: subtitles
track.label: demo label
track.language: en-GB
track.mode: hidden
addtrack event called
